requirements.txt
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
cycler==0.11.0
dask==2022.4.0
DateTime==4.4
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distributed==2022.4.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fonttools==4.30.0
fsspec==2022.3.0
HeapDict==1.0.1
idna==3.3
ipykernel==6.9.2
ipython-genutils==0.2.0
ipython==8.1.1
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.4.0
llvmlite==0.38.0
locket==0.2.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
matplotlib==3.5.1
matrixprofile==1.1.10
mistune==0.8.4
msgpack==1.0.3
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
nltk==3.7
notebook==6.4.9
numba==0.55.1
numpy==1.20.0
openpyxl==3.0.9
packaging==21.3
panda==0.3.1
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==22.0.4
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.11.2
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.3.2
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==57.0.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.1
stack-data==0.2.0
statsmodels==0.13.2
stumpy==1.11.1
tblib==1.7.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
toolz==0.11.2
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
tsfresh==0.19.0
typing-extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
widgetsnbextension==3.5.2
wordcloud==1.8.1
zict==2.1.0
zope.interface==5.4.0
